push simplification further
